refactored to comply with #9
