+++ printer/filter.py
@@ -1,5 +1,5 @@
 from forms.form import Form
-from printer.templates.mission import MissionTemplate
+from printer.templates.mission_template import MissionTemplate
 from printer.templates.template import Template
 
 from forms.mission_form import MissionForm
